changed database

+++ db/migrate/20211221054747_create_transactions.rb
@@ -2,7 +2,7 @@ class CreateTransactions < ActiveRecord::Migration[6.1]
   def change
     create_table :transactions do |t|
       t.string :type
-      t.float :amount
+      t.decimal :amount
       t.date :date
 
       t.timestamps
